:memo: [Modify] 테이블 타입 변경에 따른 코드 수정(완료), 카카오계정을 통한 신규 유저 생성(완료)

+++ src/main/java/com/team3/itability/mypage/dao/MemberProfileDAO.java
@@ -5,6 +5,6 @@ import com.team3.itability.mypage.dto.MemberProfileDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface MemberProfileDAO extends JpaRepository<MemberProfileDTO,Integer> {
+public interface MemberProfileDAO extends JpaRepository<MemberProfileDTO,Long> {
 
 }
