+++ src/main/java/com/team3/itability/mypage/dao/MemberSkillDAO.java
@@ -3,6 +3,6 @@ package com.team3.itability.mypage.dao;
 import com.team3.itability.mypage.dto.MemberSkillDTO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberSkillDAO extends JpaRepository<MemberSkillDTO,Integer> {
+public interface MemberSkillDAO extends JpaRepository<MemberSkillDTO,Long> {
 
 }
